+++ Spin-Pay/routes/web.php
@@ -3,6 +3,9 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Mailes;
 use App\Http\Controllers\AgentDashboardController;
+use App\Http\Controllers\Borrower;
+use App\Http\Controllers\Lender;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -43,13 +46,9 @@ Route::get('/register/userdocuments/{id}',function($id){
 
 
 
-// Borrower
-Route::get('/user/borrower', function(){
-    return view('user.borrower.dashboard');
-});
-Route::get('/user/lender', function(){
-    return view('user.lender.dashboard');
-});
+//Users Routes
+Route::get('/user/borrower',[Borrower::class,'borrowerdashboard']);
+Route::get('/user/lender',[Lender::class,'lenderdashboard']);
 
 
 //Agent Routes
